Support pdftkPath for no windows platform
If we specify pdftkPath and platform is not windows, it's not necessary to use execFile instead exec. I'd like to do this change because I need to specify pdftk path on linux platform as I don't install pdftk with apt-get command. This is usefull when pdftk is used on cloud platforms as AWS or Heroku.